+++ CHANGELOG.md
@@ -2,9 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## 2024-04-25
+
+### Added
+
+### Changed
+
+- Can now run without any parameters: The current directory is used as the repository path.
+- Use `--repository owner/repo` instead of `--owner` and `--repo`.
+- The defaults for `--startrevision` and `--endrevision` are now the complete branch history (starting from `HEAD`).
+
+### Fixed
+
+- Fixed typo in parameter `--login-mapping`.
+
+## 2024-04-09
 
 Initial release.